CameraHelper: panning and zooming
