+++ src/geo/projection/camera_helper.ts
@@ -1,14 +1,28 @@
 import Point from '@mapbox/point-geometry';
-import {IReadonlyTransform} from '../transform_interface';
+import {IReadonlyTransform, ITransform} from '../transform_interface';
 import {LngLat} from '../lng_lat';
+
+export type MapControlsDeltas = {
+    panDelta: Point;
+    zoomDelta: number;
+    bearingDelta: number;
+    pitchDelta: number;
+    around: Point;
+}
 
 /**
  * @internal
  * Contains projection-specific functions related to camera controls, easeTo, flyTo, inertia, etc.
  */
 export interface ICameraHelper {
+    get useGlobeControls(): boolean;
+
     handlePanInertia(pan: Point, transform: IReadonlyTransform): {
         easingCenter: LngLat;
         easingOffset: Point;
     };
+
+    handleMapControlsPitchBearingZoom(deltas: MapControlsDeltas, tr: ITransform): void;
+
+    handleMapControlsPan(deltas: MapControlsDeltas, tr: ITransform, preZoomAroundLoc: LngLat): void;
 }
