+++ src/geo/projection/mercator_camera_helper.ts
@@ -1,12 +1,14 @@
 import Point from '@mapbox/point-geometry';
 import {LngLat} from '../lng_lat';
-import {IReadonlyTransform} from '../transform_interface';
-import {ICameraHelper} from './camera_helper';
+import {IReadonlyTransform, ITransform} from '../transform_interface';
+import {ICameraHelper, MapControlsDeltas} from './camera_helper';
 
 /**
  * @internal
  */
 export class MercatorCameraHelper implements ICameraHelper {
+    get useGlobeControls(): boolean { return false; }
+
     handlePanInertia(pan: Point, transform: IReadonlyTransform): {
         easingCenter: LngLat;
         easingOffset: Point;
@@ -15,5 +17,15 @@ export class MercatorCameraHelper implements ICameraHelper {
             easingOffset: pan,
             easingCenter: transform.center,
         };
+    }
+
+    handleMapControlsPitchBearingZoom(deltas: MapControlsDeltas, tr: ITransform): void {
+        if (deltas.bearingDelta) tr.setBearing(tr.bearing + deltas.bearingDelta);
+        if (deltas.pitchDelta) tr.setPitch(tr.pitch + deltas.pitchDelta);
+        if (deltas.zoomDelta) tr.setZoom(tr.zoom + deltas.zoomDelta);
+    }
+
+    handleMapControlsPan(deltas: MapControlsDeltas, tr: ITransform, preZoomAroundLoc: LngLat): void {
+        tr.setLocationAtPoint(preZoomAroundLoc, deltas.around);
     }
 }
